requirements.txt
MDP==3.6
numpy==1.19.1
Pillow==7.2.0
PyOpenGL==3.1.5